requirements.txt
python-twitter
plotly
emoji
python-dotenv
font-fredoka-one
inky
Pillow
